SC_Project_1718-ARS/backend/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
CacheControl==0.12.3
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-cors-headers==2.1.0
djangorestframework==3.6.4
freeze==1.0.10
idna==2.6
imdbpie==4.3.0
lockfile==0.12.2
lxml==4.1.1
msgpack-python==0.4.8
numpy==1.13.1
opencv-python==3.3.0.10
panda==0.3.1
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.0
scipy==1.0.0
selenium==3.6.0
six==1.11.0
sklearn==0.0
urllib3==1.22
whitenoise==3.3.1
